Add files via upload
Write a function that takes 2 args. The 1st arg is a string with substrings separated by a delimiter character. The 2nd arg is the delimiter character. Return an array that contains 1 element. The value of that element is the 1st substring separated by the delimiter character. Don't use str.split().

For example, if the 1st arg is "buy groceries|do homework|pay bills" and the 2nd arg is "|", The returned array will have 1 element with the value "buy groceries".